bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = "ideale-ai slack bot interface"
authors = ["JeongJaeSoon <94jaesoon.jeong@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
slack-bolt = "^1.18.1"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
asyncio = "^3.4.3"
python-dotenv = "^1.0.0"
openai = "^1.3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
